added much better parsing of the xml file to XML Parser work
